src/main/webapp/app/entities/danh-muc-dich-vu/delete/danh-muc-dich-vu-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { DanhMucDichVuService } from '../service/danh-muc-dich-vu.service';

import { DanhMucDichVuDeleteDialogComponent } from './danh-muc-dich-vu-delete-dialog.component';

describe('DanhMucDichVu Management Delete Component', () => {
  let comp: DanhMucDichVuDeleteDialogComponent;
  let fixture: ComponentFixture<DanhMucDichVuDeleteDialogComponent>;
  let service: DanhMucDichVuService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [DanhMucDichVuDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(DanhMucDichVuDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(DanhMucDichVuDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DanhMucDichVuService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
